Fix compiler warnings in scripting and range_map library
range_map.h / rm_iter.h [-Wdeprecated-declarations]:
  C++17 deprecated std::allocator::{rebind,address,construct,destroy}.
  - rebind<T>::other → std::allocator_traits<A>::rebind_alloc<T>
  - alloc.address(x) → std::addressof(x)
  - alloc.construct(ptr,...) → std::allocator_traits<A>::construct(alloc,ptr,...)
  - alloc.destroy(ptr) → std::destroy_at(ptr)  (committed previously)

vm.h:741,777 [-Wsign-compare]:
  int loop counter compared against sizeof...(Args) (size_t).
  Fix: cast sizeof...(Args) to int at comparison site.

DebugExecutionManager.cpp:250 [-Wtautological-constant-out-of-range-compare]:
  EVMAbortException enum compared against size_t constant.
  Fix: cast enum to size_t at comparison site.

vmbuilder.cpp:1156 [-Winvalid-offsetof]:
  offsetof used on PFunction (non-standard-layout due to virtual methods).
  Well-defined on all target compilers; suppress locally with pragma.

+++ src/common/scripting/dap/DebugExecutionManager.cpp
@@ -247,7 +247,7 @@ void DebugExecutionManager::HandleException(EVMAbortException reason, const std:
 			event.text = message;
 		}
 		event.reason = "exception";
-		event.description = "Paused on exception: " + (reason < exceptionStringsSize ? std::string(exceptionStrings[(int)reason]) : "Unknown");
+		event.description = "Paused on exception: " + ((size_t)reason < exceptionStringsSize ? std::string(exceptionStrings[(int)reason]) : "Unknown");
 		event.threadId = 1;
 		m_session->send(event);
 	};
